feat: add cjsToEsm function
