+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,37 @@
+import * as path from 'path';
+import * as fs from 'fs';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { dirname } from 'path';
+import { fileURLToPath } from 'url';
+import './files/c.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = JSON.parse(fs.readFile(path.join(__dirname, 'files', 'a.json'), 'utf-8'));
+} else {
+    unknownObject = JSON.parse(fs.readFile(path.join(__dirname, 'files', 'b.json'), 'utf-8'));
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export default {
+    unknownObject,
+    createMyServer,
+};
+
